@@ -9,21 +9,23 @@
 
 namespace App\Controller;
 
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\RouterInterface;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class DefaultController extends AbstractController
 {
     /**
      * @Route("/", name="default")
      */
-    public function index(RouterInterface $router)
+    public function index(RouterInterface $router) :Response
     {
-        $result = array_filter(array_keys( $router->getRouteCollection()->all()), function($v){
-            return preg_match("/admin_/", $v);
+        $result = array_filter(array_keys($router->getRouteCollection()->all()), function ($v) {
+            return preg_match('/admin_/', $v);
         });
         dd($result);
+
         return $this->render('admin/default/index.html.twig', [
             'controller_name' => 'DefaultController',
         ]);
