@@ -9,15 +9,18 @@
 
 namespace App\Core\Enum;
 
+use phpDocumentor\Reflection\Types\Integer;
+use PhpParser\Node\Scalar\String_;
+
 abstract class GenreEnum
 {
-    const GENRE_MR = 0;
-    const GENRE_MME = 1;
+    const GENRE_MR   = 0;
+    const GENRE_MME  = 1;
     const GENRE_MLLE = 2;
 
-    public static $typeName = [
-        self::GENRE_MR => 'MR',
-        self::GENRE_MME => 'Mme',
+    public static array $typeName = [
+        self::GENRE_MR   => 'MR',
+        self::GENRE_MME  => 'Mme',
         self::GENRE_MLLE => 'Mlle',
     ];
 
@@ -26,7 +29,7 @@ abstract class GenreEnum
      *
      * @return string
      */
-    public static function getTypeName($typeShortName)
+    public static function getTypeName($typeShortName): string
     {
         if (!isset(static::$typeName[$typeShortName])) {
             return "Unknown type ($typeShortName)";
@@ -36,7 +39,7 @@ abstract class GenreEnum
     }
 
     /**
-     * @return array<string>
+     * @return array<int>
      */
     public static function getAvailableTypes(): array
     {
