@@ -13,7 +13,7 @@ use Symfony\Component\Security\Core\Exception\CustomUserMessageAuthenticationExc
 
  final class UserBannedException extends CustomUserMessageAuthenticationException
  {
-     public function __construct($message = 'Ce compte a été bloqué', array $messageData = [], $code = 0, Throwable $previous = null)
+     public function __construct($message = 'Ce compte a été bloqué', array $messageData = [], $code = 0, \Throwable $previous = null)
      {
          parent::__construct($message, $messageData, $code, $previous);
      }
