@@ -9,10 +9,10 @@
 
 namespace App\Domain\Membre\Entity;
 
-use App\Domain\Membre\Repository\ResetPasswordRequestRepository;
 use Doctrine\ORM\Mapping as ORM;
-use SymfonyCasts\Bundle\ResetPassword\Model\ResetPasswordRequestInterface;
+use App\Domain\Membre\Repository\ResetPasswordRequestRepository;
 use SymfonyCasts\Bundle\ResetPassword\Model\ResetPasswordRequestTrait;
+use SymfonyCasts\Bundle\ResetPassword\Model\ResetPasswordRequestInterface;
 
 /**
  * @ORM\Entity(repositoryClass=ResetPasswordRequestRepository::class)
